continue fix warning in M4

+++ src/PikaVM.c
@@ -382,17 +382,17 @@ static Arg* VM_instruction_handler_OPT(PikaObj* self,
         goto OPT_exit;
     }
     if (strEqu(">=", data)) {
-        outArg =
-            arg_setInt(outArg, "",
-                       (num1_f > num2_f) ||
-                           ((num1_f - num2_f) * (num1_f - num2_f) < 0.000001));
+        outArg = arg_setInt(
+            outArg, "",
+            (num1_f > num2_f) ||
+                ((num1_f - num2_f) * (num1_f - num2_f) < (float)0.000001));
         goto OPT_exit;
     }
     if (strEqu("<=", data)) {
-        outArg =
-            arg_setInt(outArg, "",
-                       (num1_f < num2_f) ||
-                           ((num1_f - num2_f) * (num1_f - num2_f) < 0.000001));
+        outArg = arg_setInt(
+            outArg, "",
+            (num1_f < num2_f) ||
+                ((num1_f - num2_f) * (num1_f - num2_f) < (float)0.000001));
         goto OPT_exit;
     }
     if (strEqu("&", data)) {
